fix: ensure query embedding uses float32 to match loaded embeddings and prevent dtype errors

+++ backend.py
@@ -143,7 +143,7 @@ class QueryRequest(BaseModel):
 @app.post('/search')
 async def search(request: QueryRequest):
     try:
-        query_embedding = model.encode(request.query, convert_to_tensor=True, dtype=torch.float16)
+        query_embedding = model.encode(request.query, convert_to_tensor=True, dtype=torch.float32)
         # 1. Try Q&A semantic search first (medical/nursing domain)
         if qa_embeddings is not None and len(qa_questions) > 0:
             qa_hits = util.semantic_search(query_embedding, qa_embeddings, top_k=5)
